requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
python-socketio==5.10.0
pydantic==2.5.2
asyncio==3.4.3
python-jose==3.3.0
colorama==0.4.6
flask>=2.0.0
flask-cors>=3.0.0
werkzeug==2.0.0
jinja2==3.0.3
itsdangerous==2.0.1
markupsafe==2.0.1
google-generativeai==0.3.1
aiohttp==3.8.5
python-dotenv>=0.19.0
requests>=2.26.0
google-cloud-aiplatform>=1.0.0 